requirements.txt
pandas
gradio
transformers
huggingface_hub
python-dotenv
